Port Cxtie controller to new architecture
