+++ src/controllers/cxtie/tempy-wempy.listener.ts
@@ -1,0 +1,44 @@
+import { GuildTextBasedChannel } from "discord.js";
+
+import getLogger from "../../logger";
+import {
+  CooldownManager,
+  useCooldown,
+} from "../../middleware/cooldown.middleware";
+import {
+  contentMatching,
+  isPollutionImmuneChannel,
+} from "../../middleware/filters.middleware";
+import { MessageListenerBuilder } from "../../types/listener.types";
+import { replySilently } from "../../utils/interaction.utils";
+import { formatContext } from "../../utils/logging.utils";
+
+const log = getLogger(__filename);
+
+const onTempyWempy = new MessageListenerBuilder().setId("tempy-wempy");
+
+onTempyWempy.filter(contentMatching(/tempy wempy/i));
+onTempyWempy.execute(async (message) => {
+  const channel = message.channel as GuildTextBasedChannel;
+  let reacted: boolean;
+  if (isPollutionImmuneChannel(channel)) {
+    await message.react("🇸");
+    await message.react("🇹");
+    await message.react("🇴");
+    await message.react("🇵");
+    reacted = true;
+  } else {
+    await replySilently(message, "Stop calling me that.");
+    reacted = false;
+  }
+  log.debug(
+    `${formatContext(message)}: protested being called "tempy wempy" ` +
+    `(${reacted ? "reacted" : "replied"}).`
+  );
+});
+
+const cooldown = new CooldownManager({ type: "global", seconds: 60 });
+onTempyWempy.filter(useCooldown(cooldown));
+onTempyWempy.saveCooldown(cooldown);
+
+export default onTempyWempy.toSpec();
